editor-master/requirements.txt
antlr4-python2-runtime==4.7.2
antlr4-python3-runtime==4.7.2
click==7.1.1
Flask==1.1.2
Flask-Cors==3.0.8
graphviz==0.13.2
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
six==1.14.0
Werkzeug==1.0.1
